Reasonably major update.
Added the much better ScaleDataOverall Processor.

Added a regressor which averages other regressors.

Added saving and loading Processors, as they're really a necessary part
of the DNN model. They'll be automatically saved, currently at the same
rate that the model itself is. (KNOWN ISSUE: They are currently saved
one step earlier than the model is.)

Added the ability to create DNN just from the model_dir, as the DNN
details can be inferred from the folder name. (And are probably saved in
the data anyway, to be honest, but I've not tried digging into it.)

Added the ability to recreate multiple saved DNN at once.

Added batch normalization as an option to Sequential.

Saved a particularly nice X, y datapoint sitting right on a peak. (As
X_peak, y_peak)

General code tidy-up.