Can now train with FEniCS.
Switched the multithreading from Pools to Processes and Queues, which
fit better with the way FEniCS operates. (The problem with Pools was
that pickling the data defining the FEniCS solution was too slow;
Processes side step the issue by allowing each Process to create and
maintain its own state.)

Created the necessary classes for training with FEniCS: they provide for
reusing FEniCS solutions (because they're so expensive to generate), and
for generate FEniCS solutions alongside peakon solutions, and for
feeding all of this into the DNN.

Added a couple of custom exceptions for particular errors.

Raised the logging level of FEniCS.

Started training DNN with FEniCS solutions.

(Plus all the usual minor code changes)